🚚 Move view to pages directory
