--- src/app/head.tsx
@@ -1,18 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>PokeWeb</title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <meta name="description" content="Pokedex website created with next" />
-      <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-      <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-      <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
-      <link rel="manifest" href="/site.webmanifest" />
-      <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5" />
-      <link rel="shortcut icon" href="/favicon.ico" />
-      <meta name="msapplication-TileColor" content="#da532c" />
-      <meta name="msapplication-config" content="/browserconfig.xml" />
-      <meta name="theme-color" content="#ffffff" />
-    </>
-  );
-}
