+++ src/pages/pokemons/[pokemonName]/index.tsx
