+++ src/theme/index.ts
@@ -1,0 +1,2 @@
+export { default as colors } from './theme.colors';
+export { default as theme } from './theme';
